fix: added pacman loader to dashboard so it doesn't load before it gets portfolio value from the api
